feature/#10:migration app to TS
