--- index.js
@@ -1,9 +1,0 @@
-const app = require('./src/app');
-
-require('dotenv').config();
-
-const PORT = process.env.PORT;
-console.log(PORT);
-app.listen(PORT, () => {
-    console.log('Server started');
-});
