+++ index.ts
@@ -1,0 +1,9 @@
+import { app } from './src/app';
+
+import { config } from 'dotenv';
+
+config();
+
+app.listen(process.env.PORT, () => {
+    console.log('Server started');
+});
