+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+    "watch": ["./src/"],
+    "ext": "json,html,env,ts",
+    "ignore": ["src/**/*.spec.ts"],
+    "exec": "ts-node --files ./index.ts"
+}
