+++ package.json
@@ -4,8 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start:dev": "nodemon index.js",
-    "start:devtest": "nodemon tasks.controller.js"
+    "dev:start": "nodemon"
   },
   "repository": {
     "type": "git",
@@ -22,9 +21,12 @@
     "body-parser": "^1.19.0",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "pg": "^8.7.1"
+    "pg": "^8.7.1",
+    "ts-node": "^10.2.0",
+    "typescript": "^4.3.5"
   },
   "devDependencies": {
+    "@types/node": "^16.4.13",
     "nodemon": "^2.0.12"
   }
 }
