+++ tsconfig.json
@@ -1,0 +1,17 @@
+{
+    "compilerOptions": {
+      "outDir": "dist",
+      "module": "commonjs",
+      "noImplicitAny": false,
+      "removeComments": true,
+      "preserveConstEnums": true,
+      "sourceMap": true,
+      "types": ["jest", "node"],
+      "typeRoots": ["./types", "@types", "node_modules/@types"],
+      "strictNullChecks": false,
+      "esModuleInterop": true,
+      "strict": true,
+      "resolveJsonModule": true
+    },
+    "exclude": ["node_modules"]
+}
